Move existing console app into dataconvert folder
Move existing console app into dataconvert folder
