+++ sec/consoleapp/dataconverter/conf/app.conf
@@ -1,7 +1,7 @@
-host_mssql=EACIIT-REZA
+host_mssql=ECSBYSVR01
 database_mssql=ecsec
-username_mssql=EACIIT-REZA\Wind 8
-password_mssql=ic4l
+username_mssql=sa
+password_mssql=Password.Sql
 host_mongo=127.0.0.1:27017
 database_mongo=ecsec
 username_mongo=
