+++ sec/consoleapp/dataconverter/models/scenariosimulation.go
